requirements.txt
eth_abi==4.1.0
eth_typing==3.4.0
loguru==0.7.0
python-dotenv==1.0.0
web3==6.8.0
